+++ examples/all/configs/auth.yaml
