+++ examples/auth/auth.yaml
